Timing Overhaul
- Block progress rate and block spawn rate are now locked into the tick
rate.
-- This means that when the block spawn rate changes, the tick counter
will not need to catch up to the next interval for the new specified
block spawn rate.
-- TODO Move the tick frequency changer outside the tick loop so that it
does now experience an exponential increase.
- Rotated pixel pins 180° so that device can be held from the other
side. Allows for better grip for pushing buttons, allows for higher view
of the screen.
- TODO Find the current values for the button input values (the 'biv'
array).
- TODO Find a way to separate the LED's on the hardware using paper or
cardboard so that it's more obvious which ones are lit and which aren't.